Time: 15 ms (41.98%), Space: 45 MB (10.43%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.java
@@ -7,8 +7,8 @@ class Solution {
             return true;
         }
         
-        long input = x;
-        long reverse = 0;
+        int input = x;
+        int reverse = 0;
         
         while(input>0){
             
@@ -19,7 +19,7 @@ class Solution {
         
         //System.out.println(reverse);
         
-        return ((long) x ) == reverse;
+        return x == reverse;
     }
 }
